Change _normalizeTypeKey to defer normalization to the container
There are some failing tests and some additional work to do.
The issue is that in unit tests for ember-cli for models with dashes and
when using an adapter, the model's typeKey is normalized to camelCase
but `store.modelFor(camelCasedName)` will fail.

For example, with a model class 'ssh-key', create a record using the
string 'ssh-key'. This will work fine, but later when saving that
record, the store will look up the adapter and then read off the model
class's typeKey (which is 'sshKey'), and pass it to store.modelFor which
will fail because it can't find a model called 'sshKey'.